+++ js/main.js
@@ -1,11 +1,9 @@
 import {popupShowHide} from './popup-show-hide.js';
-import {editEffects} from './edit-effects.js';
 import {editDescription} from './edit-description.js';
 import {checkData} from './check-data.js';
 import './edit-scale.js';
 import './thumbnails-show.js';
 
 popupShowHide();
-editEffects();
 editDescription();
 checkData();
